Filesys rework
