+++ retrodep/fsdb_host.c
@@ -1,0 +1,461 @@
+#include "fsdb.h"
+#include "zfile.h"
+
+#include <sys/time.h>
+#if !defined(_WIN32) && !defined(__CELLOS_LV2__)
+#include <sys/timeb.h>
+#endif
+
+#ifdef __CELLOS_LV2__
+#include "ps3_headers.h"
+#endif
+
+#ifdef VITA
+#include <psp2/types.h>
+#include <psp2/io/dirent.h>
+#include <psp2/kernel/threadmgr.h>
+#define rmdir(name) sceIoRmdir(name)
+#define mkdir(name) sceIoMkdir(name, 0777)
+#endif
+
+#if defined(__CELLOS_LV2__) || defined(_WIN32) || defined(WIIU) || defined(VITA)
+#define lstat stat
+#endif
+
+extern int log_filesys;
+
+bool my_stat (const TCHAR *name, struct mystat *ms) {
+	struct stat sonuc;
+	if (stat(name, &sonuc) == -1) {
+		write_log("my_stat: stat on file %s failed\n", name);
+		return false;
+	}
+	if (log_filesys)
+		write_log("fs_stat returned size %9jd: %s\n", sonuc.st_size, name);
+	ms->size = sonuc.st_size;
+	ms->mode = 0;
+	if (sonuc.st_mode & S_IRUSR) {
+		ms->mode |= FILEFLAG_READ;
+	}
+	if (sonuc.st_mode & S_IWUSR) {
+		ms->mode |= FILEFLAG_WRITE;
+	}
+	ms->mtime.tv_sec = sonuc.st_mtime;
+	ms->mtime.tv_usec = 0;
+#ifdef HAVE_STRUCT_STAT_ST_BLOCKS
+	ms->st_blocks = sonuc.blocks;
+#endif
+	return true;
+}
+
+static int setfiletime (const TCHAR *name, int days, int minute, int tick, int tolocal)
+{
+	return 0;
+}
+
+bool my_utime (const TCHAR *name, struct mytimeval *tv)
+{
+	int tolocal;
+	int days, mins, ticks;
+	struct mytimeval tv2;
+#if !defined(__CELLOS_LV2__) && !defined(WIIU) && !defined(__SWITCH__) && !defined(VITA)
+	if (!tv) {
+		struct timeb time;
+		ftime (&time);
+		tv2.tv_sec = time.time;
+		tv2.tv_usec = time.millitm * 1000;
+		tolocal = 0;
+	} else
+#else
+		tv2.tv_sec = 1;
+		tv2.tv_usec = 1000;
+		tolocal = 0;
+#endif
+	{
+		tv2.tv_sec = tv->tv_sec;
+		tv2.tv_usec = tv->tv_usec;
+		tolocal = 1;
+	}
+	timeval_to_amiga (&tv2, &days, &mins, &ticks);
+	if (setfiletime (name, days, mins, ticks, tolocal))
+		return true;
+
+	return false;
+}
+
+int my_existsfile (const char *name)
+{
+#if 1
+	FILE *file = fopen(name, "r");
+	if (file)
+	{
+		fclose(file);
+		return 1;
+	}
+	return 0;
+#else
+	struct stat sonuc;
+	if (lstat (name, &sonuc) == -1) {
+		return 0;
+	} else {
+		if (!S_ISDIR(sonuc.st_mode))
+			return 1;
+	}
+	return 0;
+#endif
+}
+
+int my_existsdir(const char *name)
+{
+#if 1
+	DIR *dir = opendir(name);
+	if (dir)
+	{
+		closedir(dir);
+		return 1;
+	}
+	return 0;
+#else
+	struct stat sonuc;
+
+	if (lstat (name, &sonuc) == -1) {
+		return 0;
+	} else {
+		if (S_ISDIR(sonuc.st_mode))
+			return 1;
+	}
+	return 0;
+#endif
+}
+
+int my_getvolumeinfo(const char *root)
+{
+	struct stat sonuc;
+	int ret = 0;
+
+	if (lstat(root, &sonuc) == -1)
+		return -1;
+	if (!S_ISDIR(sonuc.st_mode))
+		return -1;
+
+	ret |= MYVOLUMEINFO_STREAMS;
+	return ret;
+}
+
+FILE *my_opentext(const TCHAR *name)
+{
+	FILE *f;
+	uae_u8 tmp[4];
+	int v;
+
+	f = fopen(name, "rb");
+	if (!f)
+		return NULL;
+	v = fread(tmp, 1, 4, f);
+	fclose(f);
+	if (v == 4) {
+		if (tmp[0] == 0xef && tmp[1] == 0xbb && tmp[2] == 0xbf)
+			return fopen (name, "r, ccs=UTF-8");
+		if (tmp[0] == 0xff && tmp[1] == 0xfe)
+			return fopen (name, "r, ccs=UTF-16LE");
+	}
+	return fopen(name, "r");
+}
+
+struct my_opendir_s *my_opendir(const TCHAR *name, const TCHAR *mask)
+{
+	struct my_opendir_s *mod;
+	mod = xmalloc (struct my_opendir_s, 1);
+	if (!mod)
+		return NULL;
+
+	DIR* dir = opendir(name);
+	if (!dir) {
+		write_log("my_opendir '%s' failed\n", name);
+		return NULL;
+	} else if (log_filesys)
+		write_log("my_opendir '%s'\n", name);
+
+	mod->dh = dir;
+	return mod;
+}
+
+void my_closedir(struct my_opendir_s *mod)
+{
+	if (mod)
+	{
+		closedir(mod->dh);
+		free(mod);
+	}
+}
+
+int my_readdir(struct my_opendir_s* mod, TCHAR* name)
+{
+	mod->dp = readdir(mod->dh);
+	if (!mod->dp)
+		return 0;
+	_tcscpy (name, mod->dp->d_name);
+	if (log_filesys)
+		write_log("my_readdir => '%s'\n", name);
+	return 1;
+}
+
+int my_mkdir(const TCHAR *name)
+{
+#ifdef __WIN32__
+	return mkdir(name);
+#else
+	return mkdir(name, 0777);
+#endif
+}
+
+int my_rmdir(const TCHAR *name)
+{
+	struct my_opendir_s *od;
+	int cnt;
+	TCHAR tname[MAX_DPATH];
+	memset(tname, 0, sizeof(TCHAR) * MAX_DPATH);
+
+	/* SHFileOperation() ignores FOF_NORECURSION when deleting directories.. */
+	od = my_opendir(name, 0);
+	if (!od)
+		return -1;
+	cnt = 0;
+	while (my_readdir(od, tname)) {
+		if (!_tcscmp(tname, _T(".")) || !_tcscmp(tname, _T("..")))
+			continue;
+		cnt++;
+		break;
+	}
+	my_closedir(od);
+	if (cnt > 0)
+		return -1;
+
+	return rmdir(name);
+}
+
+int my_unlink(const TCHAR *name)
+{
+	return unlink(name);
+}
+
+int my_rename(const TCHAR *oldname, const TCHAR *newname)
+{
+	return rename(oldname, newname);
+}
+
+struct my_openfile_s *my_open(const TCHAR *name, int flags)
+{
+	if (log_filesys)
+		write_log("my_open %s flags=%x\n", name, flags);
+
+#ifdef FD_OPEN
+	int open_flags = O_BINARY;
+	if (flags & O_TRUNC)
+		open_flags = open_flags | O_TRUNC;
+	if (flags & O_CREAT)
+		open_flags = open_flags | O_CREAT;
+
+	if (flags & O_RDWR)
+		open_flags = open_flags | O_RDWR;
+	else if (flags & O_RDONLY)
+		open_flags = open_flags | O_RDONLY;
+	else if (flags & O_WRONLY)
+		open_flags = open_flags | O_WRONLY;
+
+	int fd = open(name, open_flags, 0644);
+#else
+	FILE *fp = NULL;
+	char* fopen_flags = strdup("rb");
+
+	if (flags & O_RDWR)
+		fopen_flags = strdup("rb+");
+	else if (flags & O_WRONLY)
+		fopen_flags = strdup("wb");
+
+	if (flags & O_TRUNC)
+		fopen_flags = strdup("wb+");
+	if (flags & O_CREAT)
+		fopen_flags = strdup("wb");
+
+	fp = fopen(name, fopen_flags);
+
+	free(fopen_flags);
+	fopen_flags = NULL;
+#endif
+	struct my_openfile_s *mos;
+	mos = xmalloc (struct my_openfile_s, 1);
+	if (!mos)
+		return NULL;
+
+#ifdef FD_OPEN
+	mos->fd = fd;
+#else
+	mos->fp = fp;
+#endif
+	mos->path = my_strdup(name);
+	return mos;
+}
+
+void my_close(struct my_openfile_s* mos)
+{
+	if (log_filesys)
+		write_log("my_close (%s)\n", mos->path);
+#ifdef FD_OPEN
+	int result = close(mos->fd);
+#else
+	int result = fclose(mos->fp);
+#endif
+	if (result != 0)
+		write_log("error %d closing file '%s'\n", result, mos->path);
+	free(mos->path);
+	free(mos);
+}
+
+uae_s64 my_lseek(struct my_openfile_s *mos, uae_s64 offset, int whence)
+{
+	if (log_filesys)
+		write_log("my_lseek %s %lld %d\n", mos->path, offset, whence);
+#ifdef FD_OPEN
+	off_t result = lseek(mos->fd, offset, whence);
+#else
+	off_t result = fseek(mos->fp, offset, whence);
+	result = (result == 0) ? ftell(mos->fp) : -1;
+#endif
+	if (log_filesys)
+	    write_log("my_lseek result %jd\n", result);
+	return result;
+}
+
+int my_truncate(const TCHAR *name, uae_u64 len) {
+	int int_len = (int) len;
+	if (log_filesys)
+		write_log("my_truncate %s len = %d\n", name, int_len);
+#ifdef FD_OPEN
+	struct my_openfile_s *mos = my_open(name, O_WRONLY);
+	if (mos == NULL) {
+		write_log("WARNING: opening file for truncation failed\n");
+		return -1;
+	}
+	int result = ftruncate(mos->fd, int_len);
+	my_close(mos);
+#else
+	int result = truncate(name, int_len);
+	if (result < 0)
+		write_log("Truncate failed\n");
+#endif
+	return result;
+}
+
+uae_s64 my_fsize(struct my_openfile_s* mos) {
+	struct stat sonuc;
+#ifdef FD_OPEN
+	if (fstat(mos->fd, &sonuc) == -1) {
+		write_log("my_fsize: fstat on file %s failed\n", mos->path);
+		return -1;
+	}
+#else
+	if (stat(mos->path, &sonuc) == -1) {
+		write_log("my_fsize: stat on file %s failed\n", mos->path);
+		return -1;
+	}
+#endif
+	return sonuc.st_size;
+}
+
+unsigned int my_read(struct my_openfile_s *mos, void *b, unsigned int size) {
+#ifdef FD_OPEN
+	ssize_t bytes_read = read(mos->fd, b, size);
+#else
+	ssize_t bytes_read = fread(b, 1, size, mos->fp);
+#endif
+	if (log_filesys)
+		write_log("my_read size=%d => %zd\n", size, bytes_read);
+	return (unsigned int) bytes_read;
+}
+
+unsigned int my_write(struct my_openfile_s *mos, void *b, unsigned int size) {
+#ifdef FD_OPEN
+	ssize_t bytes_written = write(mos->fd, b, size);
+#else
+	ssize_t bytes_written = fwrite(b, 1, size, mos->fp);
+#endif
+	if (log_filesys)
+		write_log("my_write buffer=%p size=%d => %zd\n", b, size, bytes_written);
+	return (unsigned int) bytes_written;
+}
+
+void my_canonicalize_path(const TCHAR *path, TCHAR *out, int size)
+{
+	_tcsncpy(out, path, size);
+	out[size - 1] = 0;
+	return;
+}
+
+int my_issamevolume(const TCHAR *path1, const TCHAR *path2, TCHAR *path)
+{
+	TCHAR p1[MAX_DPATH];
+	TCHAR p2[MAX_DPATH];
+	unsigned int len, cnt;
+
+	my_canonicalize_path(path1, p1, sizeof p1 / sizeof (TCHAR));
+	my_canonicalize_path(path2, p2, sizeof p2 / sizeof (TCHAR));
+	len = _tcslen (p1);
+	if (len > _tcslen (p2))
+		len = _tcslen (p2);
+	if (_tcsnicmp (p1, p2, len))
+		return 0;
+	_tcscpy (path, p2 + len);
+	cnt = 0;
+	for (unsigned int i = 0; i < _tcslen (path); i++) {
+		if (path[i] == '\\' || path[i] == '/') {
+			path[i] = '/';
+			cnt++;
+		}
+	}
+	if (log_filesys)
+	    write_log (_T("'%s' (%s) matched with '%s' (%s), extra = '%s'\n"), path1, p1, path2, p2, path);
+	return cnt;
+}
+
+int my_setcurrentdir(const TCHAR *curdir, TCHAR *oldcur)
+{
+#ifdef VITA
+	return 0;
+#else
+	int ret = 0;
+	if (oldcur)
+		ret = getcwd (oldcur, MAX_DPATH) ? 0 : 1;
+	if (curdir) {
+		const TCHAR *namep;
+		TCHAR path[MAX_DPATH];
+		namep = curdir;
+		ret = chdir (namep);
+	}
+#if 0
+	write_log("curdir=\"%s\" oldcur=\"%s\" ret=%d\n", curdir, oldcur, ret);
+#endif
+	return ret;
+#endif
+}
+
+bool my_resolvesoftlink(TCHAR *linkfile, int size)
+{
+	return false;
+}
+
+bool my_isfilehidden(const TCHAR *path)
+{
+	return 0;
+}
+
+void my_setfilehidden(const TCHAR *path, bool hidden)
+{
+}
+
+TCHAR *target_expand_environment(const TCHAR *path, TCHAR *out, int maxlen)
+{
+	if (!path)
+		return NULL;
+		return out;
+}
