+++ sources/src/include/sysdeps.h
@@ -398,7 +398,22 @@ extern int gui_message_multibutton (int flags, const char *format,...);
 #endif
 
 #ifndef O_BINARY
-#define O_BINARY 0
+#define O_BINARY 0x8000
+#endif
+#ifndef O_TRUNC
+#define O_TRUNC 0x0200
+#endif
+#ifndef O_CREAT
+#define O_CREAT 0x0100
+#endif
+#ifndef O_RDWR
+#define O_RDWR 0x0002
+#endif
+#ifndef O_WRONLY
+#define O_WRONLY 0x0001
+#endif
+#ifndef O_RDONLY
+#define O_RDONLY 0x0000
 #endif
 
 #ifndef STATIC_INLINE
